integral SCORE_TYPE into int64_t
This prevents integer overflow corss platforms (i.e., LP64 and ILP64)

+++ src/chinese/segmentor/implementations/agendachart/segmentor_include.h
@@ -4,7 +4,8 @@
 #include "learning/perceptron/hashmap_score.h"
 #include "pair_stream.h"
 
-typedef int SCORE_TYPE;
+#include "zinttypes.h"
+typedef int64_t SCORE_TYPE;
 
 /*===============================================================
  *
